+++ requirements.txt
@@ -1,3 +1,3 @@
 pytorch-lightning>=1.7.0
 lightning-utilities>=0.3.0, !=0.4.0  # this is needed for PL 1.7
-torchvision>=0.10.*
+torchvision>=0.10.0
